+++ src/Nightfire/Strategy.php
@@ -11,5 +11,26 @@ namespace DecodeLabs\Nightfire;
 
 interface Strategy extends TypeNameProvider
 {
-    public int $maxBlocks { get; }
+    public ?int $maxBlocks { get; }
+    public ?int $minBlocks { get; }
+
+    /**
+     * @var list<string>
+     */
+    public ?array $allowedCollections { get; }
+
+    /**
+     * @var list<string>
+     */
+    public ?array $allowedCategories { get; }
+
+    /**
+     * @var list<string>
+     */
+    public ?array $blockBlacklist { get; }
+
+    /**
+     * @var array<int,list<string>>
+     */
+    public ?array $indexBlacklist { get; }
 }